test/version1/MessageResolverV1.test.ts
const assert = require('chai').assert;

import { ConfigParams } from 'pip-services3-commons-nodex';
import { MessageResolverV1 } from '../../src/version1/MessageResolverV1';

suite('MessageResolverV1', () => {
    let resolver: MessageResolverV1;

    setup(() => {
        resolver = new MessageResolverV1();
    });

    test('Resolve hardcoded template', () => {
        resolver.configure(ConfigParams.fromTuples(
            'message_templates.template1.subject', 'Subject1',
            'message_templates.template1.text', 'Text1',
            'message_templates.template1.html', 'Html1'
        ));

        let message = resolver.resolve('template1');

        assert.isObject(message);
        assert.equal(message.subject, 'Subject1');
        assert.equal(message.text, 'Text1');
        assert.equal(message.html, 'Html1');
    });

    test('Resolve missing template', () => {
        resolver.configure(ConfigParams.fromTuples(
            'message_templates.template1.subject', 'Subject1',
            'message_templates.template1.text', 'Text1',
            'message_templates.template1.html', 'Html1'
        ));

        let message = resolver.resolve('template2');

        assert.isNull(message);
    });

    test('Resolve template by id', () => {
        resolver.configure(ConfigParams.fromTuples(
            'message_templates.template1', '123'
        ));

        let message = resolver.resolve('template1');
        
        assert.isObject(message);
        assert.equal(message.template, '123');
    });
    
});
